[WIP] Revamped ShortcutForm to make more robust
Changed the majority of the Shortcut loading code to be way more robust, and clearer to follow. The code had grown organically since the beginning of DisplayMagician, and really, really required some changes to make it all more robust and easier to make changes in the future. This is the first commit of that code so that I can ensure I have an offsite backup.

There is still a bit more of testing and bug hunting to do before I can be sure that it is all ready to release.