Added singleinstance thread safety and fixed process stopping
Added SemaphoreSlim control to make sure that only one task at a time can be run. If any other task (Display Profile change or a Game Shortcut run) is attempted, it is silently ignored by DisplayMagician. This code then means that only a single instance of DisplayMagician is loaded.

Also found a bug in the StopProcess function that was stopping processes from properly being stopped. This has now been corrected, and fixes #75