Fixing Tasks so they wait until they block
This is being done as the current UI code expects the shortcuts to block. This will be changed in the future when we move to .net6 and MAUI

Also fixed a bug I discovered with the stop program still being started if the user disabled it. And added in checking for the stop prgrams existence.