Made CurrentProfile work across instances
There is a caveat with that, that the one that is running first is the one that will generate the console output. This isn't what everyone would want unfortunately, as it may not be caught from a catch/powershell file. The 'DisplayMagician CurrentProfile' command will work perfectly if there are no other DisplayMagician's running, but because we have moved to a SIngleInstance model if there is already another DisplayMagician instance running, the instance started by the batch file will close and the batch file/powershell file won't get an answer :(.

Nothing I can do about that I'm afraid.