Fixed profile invalidity bug with RunShortcut from command.
When the RunShortcut command was run, sometimes the profile used in the shortcut would  incorrectly be shown to no be possible when it was in fact possible. This was due to a missing possibility profile refresh. This has now been added so the shortcuts run correctly from the desktoip shortcuts.