Change automatic renaming of old config files
Rewrote the automatic renaming of old config files to make it more reliable, and automatically avoid renaming the files currently in use by this version of DisplayMagician. Should make it far harder for me to incorrectly break the config files.

Also make DisplayMagician log TRACE level files by default (can be adjusted by the user if wanted). This was done to make it easier for new users to log errors when things break.