[WIP] Gathering Installed Program and UWP Program data
This is the first step towards getting automatically scanning the local computer for applications and UWP applications in order to make it easier and faster to add a new application executable to monitor.

This adds a new AppLibrary concept to DisplayMagician. The idea is that if we need to add a new application source to the list, then we can easily do it using the new AppLibrary functionality that has been added.

At present, we only have the LocalLibrary in use, which scans the local PC to detect the installed applications (those installed via Add program, and those installed via Microsoft Store). LocalLibrary then provides the list of installed applications to the main DIsplayMagician code.

The next step is to get the information returned by LocalLibrary shown to the user so that they can autopopulate the Game executable information easily and quickly.

This change also incldues some speed improvements within GameLibrary and AppLibrary which will speed up the gamelibrayr scanning. The GameLibrary was scanning the game libraries twice due to the way that it was automatically lookign for games on each invocation of a Library instance. This scanning call is now done manually to give us better control over it happening.