First working exe selector though with issues
This selects installed apps or chosen exe files from the computer, though it has a couple of issues. Firstly, if an install app was previously selected, and you revisit the page, the ILV isn't scrolled to make the selection visible. It is off the screen, so doesn't look like anything is selected.

Secondly, the default image used is a steam icon. This should be changed to a new EXE file icon to highlight what it is.

Thirdly, I need to check if the icon is always being grabbed from the exe files, as there shouldnt be any broken ones!