Final tweaks to upgrade reminders
The upgrade reminders didn't clear properly when the upgrade setting was turned off. This has now been corrected, so turning off the upgrade check resets the reminder check date to today. This means that if the user turns on the upgrade check, then it will start checking straight away for new versions on startup.