Fixed incorrect communications device audio selection
Further testing showed that the communications audio device and capture devices wouldnt be set correctly if the main audio was the same as the communicaitons devices. This has been changed to separate the logic so that they change things independently of each other. THis makes the change happen even if only one part of the audio changes are required.