Fixed some lingering audio volume reverting bug
I figured out that the volume may be incorrectly reverted back if different rollback devices were used before the shortcut was run. This would cause the user to get frustrated. So now the previous rollback volumes are applied everytime that a rollback occurs on a device. This will force the settings back to what it was, rather than hoping it was correct.