[WIP] Partway through troubleshooting UWP apps
UWP apps now kinda populate the Shortcut Form properly, but we loose their details when looking for them in the list of UWP apps if the UWP app is 'chosen' again.

May need to store some more details about the chosen app in the shortcut itself (e.g. update the shortcut format).