Fixed validity checks before saving ShortcuFormt Items
Some UI changes weren't registering as being unsaved, which meant there was the oppoerunity to lose changes. This has now been rectified in this set of changes.

Also includes some WIP changes to how Apps are recorded within the Game Shortcuts. These are a bit of ane experiment and there is a long way to go before this is ready to release.