[WIP] Working Application shortcut creation
The application shortcut creation, editing, saving and loading now all works. The next step is getting the RunShortcut to run the installed application, especially if it is a UWP program!

We then need to test the 'Choose EXE from your computer' functionality to make sure the previous method for selecting programs works ok.