Upgrading everyone to v2.4.1.14
This fixes a long standing issue caused when swapping from NVIDIA Surround to a non-surround profile. DisplayConfig was causing an issue when returning on a system that uses a non-NVIDIA GPU on aa secondary screen, as the NVIDIA driver automatically disables non-NVIDIA connected screens when turning off Surround :(. Because the screen layout is changed by NVIDIA, the DisplayConfig layout is wrong, and therefore DisplayConfig fails.

The workaround was to just skip setting DisplayConfig in that instance, as WinLibrary will set 99% of the displayLayout, and as that is done through the Windows CCD interface it all works fine.

Another option to investigate later to to move the DisplayConfig layout changes to the ApplyConfigOverride function so that they are performed AFTER WinLibrary has set the correct display layout. This could work, and would then apply any special settings that are only available in DisplayConfig.