[WIP] Partially working LocalApp process monitoring
The game starting logic works, but the game detection processing doesn't.

I've also discovered that I need to refactor the process tracking logic so that it is within the libraries themselves. At the moment the logic is within RunShortcut, and that is wrong. Each App/Game/executable should be able to start, stop, and detect if it is running itself. The GameLibrary should start and stop, and the game/app/exe should know if it is running, and should be able to get it's own PID.