Updated FOV calc to remember values
Have engineered the FOV calculator to remember the settings from last time, and reuse those, but with a little twist. It also looks to see if you are in single screen, or triple screen, and changes the settings based on that. The aim is to provide the calculation that makes the most sense to the user based on the currently running screen layout.