Fully working UWP application lifecycle detection.
IMPORTANT: There is a 10 second delay between when a user closes a UWP App, and when Windows actually terminates the UWP app processes. This delay is because of the UWP application lifecycle, which is different to desktop applications lifecycle. The 10 seconds delay is in Windows, and therefore DisplayMagician can ony detect the UWP App status change once Windwos waits the 10 seconds.