Slightly improved the uninstall registry removal process
Had a report from a use that DisplayMagician doesn't correctly remove all the registry keys and files. Some of these are related to the AUMID registration that used to happen using ToastNotificationManagerCompat, so now that we've stopped using that we hopefully will avoid those issues.

Have to test it on a brand new machine to see what residual things are left behind.