Fix unsaved state when profile doesn't exist
ShortcutForm now correctly sets the UI and the file names when the display profile used in an existing shortcut no longer exists. It now correctly prompts you on close as well if you haven't saved it.

Also fixed the ShortcutForm closing when you want to save on close but the shortcut is invalid. It now stops closing the ShortcutForm, and allows you to fix the problems. Also tells you what those problems are now.