Improved logging and made Steam parsing robust
Changed to use the NLog archived file rotation so that I can use the file size attribute. Fixes an issue where DisplayMagician is running for a long period of time without the application being rebooted (I once made a 1.9GB file!)

Also added a try/catch into Steam library parsing so that if there is an issue parsing the Steam VDF files it won't creash that application. Partially fixes #173