First attempt at Steam Beta parsing
Attempts to fix #173. This adds a new Steam Magic number for their Steam API v020 of 0x07_56_44_28 (to go with the previous magic value of 0x07_56_44_27. This *may* work, but there is no guarantee. There is a new Valve ValueKeyValue file parsing library version available, but we cannot change to it until we migrate from .netFramework 4.8 to a .net6 app... and this will take a long time to do.