Added warning for UWP monitoring restrictions
Partially mitigates #98. This warning will set the users expectations for the lack of ability for DM to monitor UWP apps. Windows restricts access to the new Windows Runtime API to programs that have a PackageIdentity, and DisplayMagician doesn't have a packageidentity because its installed as a MSI, not as an MSIX. There will be quite a bit of work to get that working, so I'm going to do that as DisplayMagician v3 (where I migrate to .net6 and MAUI user interface)