Update README
Modified the README to remove the bit about selectingh the game shortcut from the backgfround. This was removed because we couldn't do it successfully for a while. This feature won't be added back in v2.x, but I will see if it can be added in DisplayMagician v3.0 which is just starting development now.