[WIP] Fixed MainScreen taskbar detection
It was incorrectly looking for the UID format in the StructRect key, which uses the 'Settings' name instead. Fixed Taskbar location detection for the single screen.

Also managed to tweak Profile detection to handle AMD laptops with NVIDIA added videocards. It now checks if the NVIDIA card actually has any displays attached, and if not ti then tries the AMD GPU mode, and if it cannot find an AMD GOPU then it defaults to Windows mode. This should fix gaming laptops screen detection logic.