Update button first attempt
This set of changes adds an update button that allows a user to update the display profile with the current display layout without needing to create it again. The profile retains the came UUID, which means that the ggame shortcuts will work without any changes. This saves users a TON of work if they need to recreate their profiles.

Additionally, this also fixes the Update button so it shows correctly.

And lastly it changes the way that the CreateProfileFromCurrentSettings fuinction works. The function now tries to figure out it's own settings as to which library it should be using, and attempts to take into account things like how many screens are connected, which video card is in charge and the like. This should make DisplayMagician much more reliable on gaming laptops or with external video cards. It *should* work better now and make more sensible decisions.