Update button now saves too!
Fixed a bug where the Update button was making the changes to the running code, but it didn't save the updated profilee to JSON. It does now.

Also disabled the Apply button once someone does an update, as updating takes the currently running settings and uses those. There is no sense to apply settings you are already using.