Validate display profile json file on save
DisplayMagician now validates the ProfileRepository JSON file when it is saved, and if there is an issue it tries to save the file again. This is to minimise the chances of any errors during the saving process, as that will completely ruin the existing DisplayProfiles.

Also improved the DisplayMagician log categorisation for some of the items that are triggered on Gaming laptops with AMD CPU with embedded GPU, and NVIDIA discrete GPU. It shouldn't compain unless it needs to :)