Remove CurrentProfile from main DisplayMagician
Due to DisplayMagician being a WinForm application it is very difficult to get it to output commands that can be used in other batch scripts. I attempted this recently and all it did was create memory leaks and delays.

My new plan is to separate out the batch-friendly commands into a specific commandline application designed for use in batch file/powershell situations.