Fixed Update button
Update button now correctly checks if the new updated profile is possible and updates the UI reflecting that fact. Previously it incorrectly left the old ispossible settings (which were just overwritten and so were now incorrect).

I also wrapped the update button with a double-check as it is somethign that can permanently affect the Display Proifile. Hopefully this avoids someone accidentally clicking the Update button!