Fixed SetActive to work with all three video libraries.
Rewrote the ProfileItem.SetActive to work with the 3 video libraries so that it will attempt to apply all the libraries correctly. This now works based on my local testing, but I will need more people to test in order to find any other hardware specific bugs.

Improved the NVIDILibrary config comparison by writing a dictionary-based comparison function that doesn't mind what the order of the dictionary or list is. This allows the config conparison to be more accommodating.