Avoid null exception error with new file
Upon first start the right-click startmenu icon logic attempted to use the CurrentProfile profile name , but that was set to null, causing a null exception. This has been fixed by creating a default name when the current profile is generated. This wasn't previously being done, which caused this issue.

Also removed the AMD/NVIDIA/Windows logos from the DisplayProfile screen so that it doesn't cover the view of the display layout.