Fixes to profile possibility detection
Due to the new modeless way that we use the videocard libraries (we use all of them at once) I have had to drastically change the detection logic that checks whether a profile is possible to use.

The logic now saves a new 'reduced' list of current displays in use, and grabs the driver information for the display from the most applicable soiurce e.g. NVIDIA connected displays get the NVIDIA device identifier, AMD connected displays get the AMD device identifier, and lastly Windows plugs the gaps. This ensures that we get one displayidentifier for each display in use. This is saved with each profile.

Additionally, I had to change the detection of displays that are connected but not in use. AMD especially had to have some tweaks to allow this to occur. Now, the allconnecteddisplays logic looks for as many displays it can find, to give us the widest range of displays to match.

The IsPossibleRefresh logic actually does the comparison checks and it checks to make sure that the profile display identifiers are all in the list of allconnecteddisplays. If they are, then we're able to use the profile!

This check happens for every profile so that we know that the profile is likely to work and so we allows the user to apply it!