Temporarily disable AMD taskbar detection
The AMD taskbar detection is really unreliable and keeps putting the taskbar on the righ tedge of the screen when drawing the display layout. This is not correct as the taskbar in reality is on the bottom.

I have been unable to change this, so I am temporarily disabling it until I get some time to work on the taskbar detection logic.