Moved to dynamic rolling archives
The last changes I made to log file archives broke the archiving function. So this time I have moved to dynamic NLog archiving, and it works - keeping 5 log files in total. This will be good for testing.

Also had to update the Support ZIP File function in the Settings UI so that it includes the new style log files.

Additionally I stopped the console log function as well, as WinForms don't have a console to log to!