Fixed AMD ADL initialisation
The AMD ADL library initialisation detection wasn't working properly, so it would keep trying to use this even if it wasn't initialised. This was throwing exceptions. This has been quelled, so hopefully it will just do nothing.

I also found the same thing in the NVIDIALibrary as well. This now correctly sets the IsInstalled variable and just logs the issue instead of throwing an exception. Hopefully this makes everything run properly again!