WinLibrary tweaks to make it more reliable
Windows in some cases wouldn't accept the WinLibrary configuration for the CCD API. It would sometimes complain of an invalid parameter. In general the same config is also installable via the GDI API (also in Winlibrary) so WinLibrary will now attempt to use the GDI settings instead when there is the parameter problem. This should make WinLibrary more reliable.

I also go rid of a lot of code that had been commented out previously as it wasn't needed any longer.