Added checks for missing video cards
Added some checks to GetAMDDisplayConfig and GetNVIDIADisplayConfig that check for the driver DLLs being installed, but the video card being unavilable. This can happen if someone changes their video card, or if the video card fails, or if the PC is a gaing laptop with NVIDIA Optimus mode. This will hopefully result in DM working properly in these situations.