Added missing videocard checks to AMD and NVIDIA Display Identifier creation functions
AMD and NVIDIA Display Identifier creation functions bothe blindly assumed there would be a videocard in there, and just operated with that. This could cause issues when there was a videocard DLL installed, but no videocard. Additionally AMD returns all videocard info (not just AMS videocards) so we had to work around that.

Both AMDLibrary and NVIDIALibrary should now avoid causing an excepotion when they are tried without the relevant videocard in the PC. This should allow the new method of using all videocard libraries at once to work.