Wrapped Video Library interaction in try/catch protection
Added try/catch protection to any VideoLibrary interaction to enable logging to catch it and rtecord it as close to source as possible. This should help me catch more video library problems easier.

Also attempted to keep DM running when it hits one of these problems so that the user may be able to get to the Settings page and generate a Support ZIP File easier even when there is a video library problem.