Disable applying GDI settings when changing layouts
I have a suspicion that applying the GDI settings then applying CCD settings is causing some of the problems where the primary display isn't always being set. It may be being caused by setting the Windows settings twice, so I've reverted to only using the modern CCD interface for this.

Additionally, I've temnporarily disabled the NVIDIA NvAPI_DISP_SetDisplayConfig functionality as it currently doesn't work anyway. There is a bug in the way that the struct handles it when the C++ DLL returns a null in the NV_DISPLAYCONFIG_PATH_TARGET_INFO_V2_INTERNAL.Details field. Structs are not nullable in C# so I need to rewrite the NV_DISPLAYCONFIG_PATH_TARGET_INFO_V2_INTERNAL and NV_DISPLAYCONFIG_PATH_TARGET_INFO_V2 Structs as classes.