Fixed DisplayMagicianConsole bugs
Tweaked and tuned the DIsplayMagicianConsole.exe so that you can use either the Profile UUID or the Profile Name to select it. Also updated the CurrentProfile command so that it reports the currentProfile UUID as well as the name, so to make it easier for  the user to use.

Also added in the ability to run "DIsplayMagicianConsole.exe AllProfiles" to get a list of all the DisplayProfiles currently available.

Also added a -p flag to change the ourput from human readable to one delimited by the vertical pipe character "|" to make it easier to parse in scripts.

Fixes #223