Removed warning meessage for UWP monitoring
It appears we may not need PackageIdentity as some testing has shown that DM can now access the AppInfo structures that allow us to monitor UWP apps. This means we should be able to track those apps and monitor them as normal.

This change disables the warning that I added before to tell people about this limitation.