Fix FOV Calculator broken RTF format output
THere was an issue with the RTF text file formatting that I'd created in previous versions. This formatting was now corrupted, and was causing the RTFTextBox to raise an exception. Rather than fix this using another RTF library creating yet another dependency, I thought it was easier to just move to a straight text input. This change avoided adding another library, but also ensurs that the output is as accessible as possible.