[WIP] Partially ot non-steam game details
Have grabbed the correct details from shortcuts.vdf and am able to grab the game appid, gamename and gamexe location.

The problem is that steam.exe -applaunch <appid> doesn't work for non-steam games. There appears to be a translation that happens with shortcuts.

The non-steam games have a negative steam appid id e.g. putty.exe has appid "-1132504025", but the steam UI generates a shortcut of  "steam://rungameid/13582676323779739648". I have to find how this is mapped.