[WIP] Additional .net8 setup installer tweaks
I still have quite a few things to fix. The installer will install the files correctly, but does not actually create a shortcut that links to the right file (it tries a dll!)

Additionally, it doesn't currently support installing the Shell Extension at all. This is something that may not be able to be done within this type of setup file, but I am hoping so.