Added per install random GUID to allow future stats
Added a random GUID identifier to each install so that in the future we would be able to gather statistics about the number of installs and possibly the O/S that they use, and the type of video card they use. These stats would be invaluable in figuring out where development time should be spent. For example I am currently figuring out whether I should add Intel Arc GPU support, and actually knowing how many people are using Intel ARc machines would help me prioritise this.

It is EXTREMELY important that we avoid any method that identifies any people actually using the software... the goal of this functionality is not to track people, but to track installs, and to provide information to direct the future development of DisplayMagician.