Partially working setup installer
This installation now registers and unregisters the DisplayMagicianShellExtension correctly, but the DisplayMagicianShellExtension itself doesn't work as it complains about the lack of .Net Framework. I think I will need to recompile SharpShell.dll to .net 8.0, or I need to find a new way to do this.

Additionally, I need to change the main text in the setup project to white.