Removal of IContextMenu Shell Extension
With the move from .Net Framework 4.8.1 to .Net 8.0 for the underlying version of .net that DisplayMagician uses, there was a conflict with the SharpShell library that we used to create teh Desktop Background menu. SharpShell requires .Net Framework 4 to work, which meant that it wouldn't work for DisplayMagician any longer. A Sharpshell .net Core version was under development but has now stalled.

This means we need an alternative. I looked around for various solutions, Looking at both IExplorerCommand and IContextMenu options, but there are no easily usable libraries for creating this functionality available within .Net 8.0. So I looked for more alternatives.

Instead, I found that Windows 7+ has the ability to do static cascading menus using registry keys, as is shown here: https://io-repo.blogspot.com/2011/05/cascading-context-menus-via-static.html.

I tested this functionality, and it works, so I can now start to change DisplayMagician to create the registry keys automatically as it runs. This new functionality means that I no longer need to force a restart after installing or uninstalling DisplayMagician.

In addition to the changes above, I have also renamed the two DisplayMagicianSetup projects to make the older WIXv4 based on different to the VS Setup Project based one I'm currently attempting to use. There are  few things that need changing like the colour of the heading text.

Lastly there are a few small changes included to correct some deprecated features that are not longer recommended, like replacing WebClient with HttpClient.