[WIP] Initial tests for ContextMenu
This contextMenu code requires admin priv to update the profile list registry keys, which means that it would require the user to give DisplayMagician admin priv each time it runs. This is not good practice and wilkl impact the usability.

I need to see if there is an application manifest setting I can set that will allow this to be done.