Initial Wix installer version
Basic setup of a new Wixv4 installer. Using the WixUI_InstallDir dialog set, but have customised the text colour and the like. Also have it set to detect at least .net 6.0.0 or higher.

I still have yet to set up the file components from the various applicatiuons. I need to create a powershell to use the heat application to scan for the files as part of the build process.

I also need to setup the installer to include the version number in the name of the output .msi file.