Added reminder to fix some code
The AppLibrary.LoadAppsInBackground() code was crashing on startup when the app was installed (it was fine in VisualStudio, and when run from the build folder). This only appreards to be a problem when the app is installed under Program Files, so I think there may be some permissions problem that is not being handled by the logic. Need to add many more try/catch statements to make this part more robust.