Fixed help URLs and PackageVersion setting
The Help URLs were using the old way of opening the setting URLs, which was causing exceptions. Fixed.

The Packageveresion was being set too early in the build process, so that package was picking up the wrong version from displaymagician.dll. Set it later in the build process to get the correct 'updated' version from the dll.