Fix DisplayMagicianConsole versioning
Fixed DisplayMagicianConsole versioning so it matches the main DisplayMagician version. Also fixed the DMPackage project so that it grabbed the published single file version of the console.exe rather than the expanded-file built version. Fixes #223.

Changed DIsplayMagicianBundle to stop attempting to change the filename as we are currently unable to extract the version information from the package for the filename. We'll worry about that later.